factor out rego package
